TP2/requirements.txt
playwright==1.55.0
aiohttp==3.13.1
beautifulsoup4==4.14.2
lxml==4.6.3
pillow==12.0.0
requests==2.32.5
redis==7.0.1